throw nice error if directory doesn't exist

+++ papermill/iorw.py
@@ -11,7 +11,7 @@ import requests
 import yaml
 
 from . import __version__
-from .exceptions import PapermillException
+from .exceptions import PapermillException, FileNotFoundError
 from .s3 import S3
 from .adl import ADL
 
@@ -88,6 +88,8 @@ class LocalHandler(object):
 
     @classmethod
     def write(cls, buf, path):
+        if not os.path.exists(path):
+            raise FileNotFoundError('output folder {} doesn\'t exist!'.format(path))
         with io.open(path, 'w', encoding="utf-8") as f:
             f.write(buf)
